Move Conference enum into the model package
