+++ src/main/java/com/gluonhq/otn/model/Service.java
@@ -25,7 +25,6 @@
  */
 package com.gluonhq.otn.model;
 
-import com.gluonhq.otn.views.ActivityFeedPresenter;
 import javafx.beans.property.ReadOnlyListProperty;
 import javafx.beans.property.ReadOnlyObjectProperty;
 import javafx.collections.ObservableList;
@@ -246,6 +245,6 @@ public interface Service {
      * Sets the selected conference.
      * @param selectedItem the selected conference
      */
-    public void setConference(ActivityFeedPresenter.Conference selectedItem);
+    public void setConference(Conference selectedItem);
     
 }
